\name{normalp-package}
\alias{normalp-package}
\docType{package}
\title{
Package for exponential power distributions (EPD)
}
\description{
  This package implements a collection of utilities referred to exponential power distributions, 
  also known as General Error Distribution. These utilities have been developed by some researcher
  of the University of Palermo, Italy.
}
\details{
\tabular{ll}{
Package: \tab normalp\cr
Type: \tab Package\cr
Version: \tab 0.7.1\cr
Date: \tab 2020-02-13\cr
License: \tab GPL
}
}
\author{
Angelo M. Mineo\cr
Maintainer: Angelo M. Mineo  <angelo.mineo@unipa.it>
}
\references{Chiodi, M. (1986) \emph{Procedures for generating pseudo-random numbers from a normal distribution of order p},
Rivista di Statistica Applicata, 19, pp. 7-26.

Mineo, A. (1989) \emph{The norm-p estimation of location, scale and simple linear regressione}, 
Lectures Notes in Statistics: Statistical Modelling, 57, pp. 222-233.

Mineo, A.M. (1994) \emph{Un nuovo metodo di stima di p per una corretta valutazione dei
parametri di intensita e di scala di una curva normale di ordine p}, 
Atti della XXXVII Riunione Scientifica della Societa Italiana di Statistica, San Remo, Vol. 2, pp. 147-154.

Mineo, A.M. (2003) \emph{On the Estimation of the Structure Parameter of a Normal Distribution of Order p}, \emph{Statistica}, anno LXIII, n. 1, pp. 109-122. 
}
\keyword{distributions}
\keyword{regression}
\keyword{package}
